chore: move requestHandler to own file
